+++ src/effects/specialFeats.js
@@ -65,6 +65,7 @@ import { deftStrikeEffect } from "./feats/deftStike.js";
 import { hadozeDodgeEffect } from "./feats/hadozeeDodge.js";
 import { mindLinkEffect } from "./feats/mindLink.js";
 import { holdBreathEffect } from "./feats/holdBreath.js";
+import { maskOfTheWildEffect } from "./feats/maskOfTheWild.js";
 
 export function baseFeatEffect(document, label) {
   let effect = {
@@ -250,6 +251,10 @@ export async function featureEffectAdjustment(ddb, character, document) {
       }
       case "Mantle of Inspiration": {
         document = await mantleOfInspirationEffect(document);
+        break;
+      }
+      case "Mask of the Wild": {
+        document = await maskOfTheWildEffect(document);
         break;
       }
       case "Patient Defense": {