+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - Sleep Spell Effect improvements for elfs and half-elfs
 - Stone of Good Luck no longer needs a initiative boosting effect separately to checks.
+- Mask of the Wild effect: PC's and NPC's will now roll stealth if this ability is clicked. (Thanks @MotoMoto)
+- Rahadin Deathly Choir effect exclusions (Thanks MotoMoto)
 
 # 3.5.8